Fix sign related issues.
Altitude and decimal degrees positions can be negative!
Without this patch, dangerous overflows can happen.
Also fix some minor documentation issues.


git-svn-id: https://src.develer.com/svnoss/bertos/trunk@3896 38d2e660-2303-0410-9eaa-f027e97ec537

+++ bertos/net/nmea.c
@@ -165,9 +165,9 @@ static udegree_t nmea_longitude(const char *plot, const char *phem)
  * Return altitude in meter from a string.
  *
  */
-static uint16_t nmea_altitude(const char *palt, const char *punits)
+static int32_t nmea_altitude(const char *palt, const char *punits)
 {
-	uint32_t alt;
+	int32_t alt;
 
 	if (*palt == 0)
 		return 0;
